+++ nutree/node.py
@@ -303,7 +303,7 @@ class Node(Generic[TData]):
         """Set `self.data` to a new string (assuming plain string node)."""
         if isinstance(self._data, str):
             return self.set_data(new_name)
-        raise ValueError("Can only rename plain string nodes")
+        raise TypeError("Can only rename plain string nodes")
 
     def set_data(
         self,
@@ -972,7 +972,7 @@ class Node(Generic[TData]):
                     p = _create_parents()
                     p._add_from(n)
                 else:
-                    raise ValueError(f"Invalid predicate return value: {res}")
+                    raise TypeError(f"Invalid predicate return value type: {res}")
 
                 parent_stack.pop()
             return
@@ -989,7 +989,7 @@ class Node(Generic[TData]):
         See also :ref:`iteration-callbacks`.
         """
         if not predicate:
-            raise ValueError("Predicate is required (use copy() instead)")
+            raise TypeError("Predicate is required (use copy() instead)")
         return self.copy(add_self=True, predicate=predicate)
 
     def filter(self, predicate: PredicateCallbackType) -> None:
@@ -998,7 +998,7 @@ class Node(Generic[TData]):
         See also :ref:`iteration-callbacks`.
         """
         if not predicate:
-            raise ValueError("Predicate is required (use copy() instead)")
+            raise TypeError("Predicate is required (use copy() instead)")
 
         def _visit(parent: Self) -> bool:
             """Return True if any descendant returned True."""